requirements.txt
flask
python-dotenv
transformers
torch
line-bot-sdk
